Reorganized test setup code
